refactored larger chunks of code into lib/*
